daemon/requirements.txt
pyrogram
tgcrypto
apscheduler
python-dotenv
httpx
bs4
import-expression
